fix(prefabs): apply slight position offset on force snap
There's an issue where if the GameObject being snapped and the
SnapZone are in the same location then a force snap is called then
the internal Object Follower won't think the position has changed
initially so will not perform the follow. This fix just applies
a slight position offset to the object being snapped just before
it is force snapped so this discrepancy is not an issue.