feat(Utility): add prefab creator
The latest version of Zinnia has the basis of a prefab creator that can
be used to enable easy adding of prefabs to a scene without needing to
drag and drop from directories. Instead a new menu item is added for
quickly adding prefabs. The guide has been updated to accommodate this
and the FodyWeavers.xml is now located in the root to serve both the
Runtime and Editor scripts.