feat(SnapZoneManager): add component to manage scene snap zones
The SnapZoneManager component provides an easy way to manage
scene snap zones. This first release only has a couple of useful
methods, one to determine if a given Interactable is in a SnapZone
(and optionally return the snap zone it is in), and to unsnap
an Interactable from a SnapZone when only the Interactable is
known.