fix(Configurator): ensure interactable rigidbody kinematic is prepared
The changes to Unity and PhysX cause an issue where the rigidbody
kinematic state change causes the collision detection to call exit
and enter again in the same frame which means the snap zone exit
and enter is called when an object is dropped into the zone even though
the object never exited the snap zone.

A solution in the Zinnia Collision Tracker was added that allows a
rigidbody to be marked for kinematic state change to overcome the issue
at hand. The snap zone now utilizes the additional events on the
interactable to know when the kinematic state changes so it can prepare
the snap zone Collision Tracker to ensure it doesn't capture the
additional enter/exit events when the interactable kinematic state
changes.