fix(Configurator): ensure prepare kinematic change is called everywhere
The PrepareKinematicStateChange mechanism was only being called when
the Interactable was directly placed within the snap zone, if it was
dropped from above and the auto snap was on then the logic would never
be called and therefore the double exit/enter would occur.

This fixes it by providing more methods for preparing the state change
with different object types just for completeness and then the prefab
now ensures before the rigidbody state is changed on the interactable
that it prepares the collider tracker for this state change.