feat(Facade): add option for highlight always on and scale optional
The Facade now has the ability to set the highlight object as always
on when the snap zone is not being used or hovered over. There is
also another option that allows the scaling properties of the snap
zone to be optional so it doesnt change the size of the snapped object
upon being snapped to the zone.